Add helper classes for query execution and SQL building
